+++ src/routes/AppRouter.jsx
@@ -5,11 +5,11 @@ import LoginPage from '../pages/LoginPage/LoginPage';
 import SignupPage from '../pages/SignupPage/SignupPage';
 import ProfileSettingPage from '../pages/ProfileSettingPage/ProfileSettingPage';
 import FeedPage from '../pages/FeedPage/FeedPage';
-import FeedDetail from '../pages/FeedPage/FeedDetail';
+import PostPage from '../pages/PostPage/PostPage';
+import PostDetailPage from '../pages/PostPage/PostDetailPage';
 import ProfilePage from '../pages/ProfilePage/ProfilePage';
 import FollowerListPage from '../pages/FollowListPage/FollowerListPage';
 import FollowingListPage from '../pages/FollowListPage/FollowingListPage';
-import PostPage from '../pages/PostPage/PostPage';
 import UploadPage from '../pages/UploadPage/UploadPage';
 import UploadMap from '../pages/UploadPage/UploadMap';
 import SearchPage from '../pages/SearchPage/SearchPage';
@@ -26,14 +26,15 @@ const AppRouter = () => {
         <Route path="/signup" element={<SignupPage />} />
         <Route path="/signup/profile" element={<ProfileSettingPage />} />
         <Route path="/home" element={<FeedPage />} />
-        <Route path="/post/:account/detail" element={<FeedDetail />} />
-        <Route path="/profile/:account" element={<ProfilePage />} />
+        {/* Post */}
+        <Route path="/post" element={<PostPage />} />
+        <Route path="/post/:id" element={<PostDetailPage />} />
+        <Route path="/post/upload" element={<UploadPage />} />
+        <Route path="/post/upload/map" element={<UploadMap />} />
+        {/* Profile */}
          <Route path="/profile/:id" element={<ProfilePage />} />
         <Route path="/profile/:id/follower" element={<FollowerListPage />} />
         <Route path="/profile/:id/following" element={<FollowingListPage />} />
-        <Route path="/post" element={<PostPage />} />
-        <Route path="/post/upload" element={<UploadPage />} />
-        <Route path="/post/upload/map" element={<UploadMap />} />
         <Route path="/search" element={<SearchPage />} />
         <Route path="/chat" element={<ChatListPage />} />
         <Route path="/chat/:id" element={<ChatRoomPage />} />
